Ran php artisan make:controller UserController --model=User --resource
User and Profile are separate things. The site has users and the users have profiles, so that's why this is needed. Breeze did not include a mechanism for seeing data related to other users, it only included mechanisms for a user seeing their own profile. We need to be able to see all Projects and Tasks assigned to various Users, so we need this User Controller.